requirements.dev.txt
# Testing
pytest
pytest-aiohttp
factory_boy

# Lint
flake8
mypy

# Migrations
alembic
psycopg2
